refactor file structure for optimal scss usage
