+++ src/App/App.js
@@ -1,9 +1,10 @@
 import React, { useState, useEffect, useContext } from 'react';
 import PropTypes from 'prop-types';
 import { Route, Link } from 'react-router-dom';
-import SearchForm from '../SearchForm/SearchForm';
+import SearchForm from '../components/SearchForm/SearchForm';
 import JointContainer from '../JointContainer/JointContainer';
-import FavPage from '../FavPage/FavPage';
+import FavPage from '../components/FavPage/FavPage';
+import Header from '../components/Header/Header';
 import './App.scss';
 import { getCityID, getJoints } from '../apiCalls.js';
 
@@ -51,13 +52,11 @@ const App = () => {
     return (
       <main className='main-section'>
         <Route exact path='/'>
-          <header >
-            <h1 className='title'>Find Some BBQ</h1>
-          </header>
+          
+          <Header />
+          
           <section>
-            <SearchForm searchCity={searchCity} />
-            
-            <JointContainer joints={joints}/>
+           
           </section>
         </Route>
