--- src/SearchForm/SearchForm.scss
@@ -1,28 +1,0 @@
-.search-form {
-  font-family: 'Raleway', sans-serif;
-  display: flex;
-  justify-content: center;
-  padding-top: 50px;
-  padding-bottom: 30px;
-  height: 36px;
-}
-
-.search-form input {
-  margin-right: 35px;
-}
-
-.search-form label {
-  margin-right: 15px;
-  align-items: center;
-  font-size: 2em;
-}
-
-.button {
-  font-family: 'Raleway', sans-serif;
-  font-size: 1.2em;
-  background-color: #f2ce4c;
-}
-/*
-font-family: 'Oswald', sans-serif;
-font-family: 'Raleway', sans-serif;
-font-family: 'Rye', cursive; */
