+++ src/components/FavPage/FavPage.scss
