+++ src/components/Favorite/Favorite.scss
