+++ src/components/Favorite/Favorite.test.js
