+++ src/components/Header/Header.scss
@@ -1,0 +1,6 @@
+
+
+
+.header {
+   
+}
