+++ src/components/Joint/Joint.scss
