+++ src/components/JointContainer/JointContainer.scss
