+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom'
 import { BrowserRouter } from 'react-router-dom';
-import './index.css';
+import './index.scss';
 // import * as serviceWorker from './serviceWorker';
 import App from './App/App';
