+++ src/styles/variables.scss
@@ -1,0 +1,10 @@
+// COLOR VARIABLES
+$color-red-dark: #db5a42; 
+$color-red-light: #e3a587;
+$color-brown: #a57f60;
+$color-blue: #20a39e;
+$color-yellow: #ffd275;
+
+// FONT VARIABLES
+$font-primary: 'Rye', cursive;
+$font-secondary: 'Raleway', sans-serif;
